fix: superfluous deep env conflicts with non-dict model leaf
If a model defines a leaf as e.g. a string and the user or system sets
an environment var with an additional `env_nested_delimiter + suffix`,
`explode_env_vars` was treating the string as a `dict` (trying to add an
item to it).

This fix checks if the target is indeed a `dict`.

In that regard, the whole `explode_env_vars` seems a bit wasteful, in
terms of possible throwing away already computed values, if a deeper
environment variable is processed earlier (which is possible, since the
iteration order of `os.environ` seems unpredictable).

resolves #275